+++ source/Private/Get-LaunchTreeDefaultConfiguration.ps1
@@ -35,7 +35,7 @@ function Get-LaunchTreeDefaultConfiguration {
         PersonalRoot     = Join-Path -Path $roamingRoot -ChildPath 'LaunchTree'
         MaximumDepth     = 5
         LauncherHost     = 'WindowsPowerShell'
-        LauncherLayout   = 'Grid'
+        LauncherLayout   = 'TabbedList'
         SortOrder        = 'NameAscending'
         CloseAfterLaunch = $true
         Cache            = [ordered] @{